requirements.txt
flask==3.0.3
requests==2.32.3
twilio==9.3.6
python-dotenv==1.0.1
